Rename aux directory into auxiliary
aux is a reserved name on Windows platforms.

This fixes #7.
